+++ Code/SX70z_ESP/SX70_ModelZ/src/display_manager.h
@@ -13,4 +13,4 @@ void display_show_taking(const camera_state_t *state, ssd1306_t *disp);
 void display_show_menu(const camera_state_t *state, ssd1306_t *disp);
 
 /** 倒计时专用显示（清屏后只显示剩余秒数） */
-void display_show_countdown(ssd1306_t *disp, int8_t seconds_remaining);
+void display_show_countdown(ssd1306_t *disp, int16_t seconds_remaining);
